feat: restructured application from pkg to blueprints
Moved configuration into config file
creation of app into a function
