+++ flaskr/__init__.py
@@ -3,21 +3,34 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.orm import DeclarativeBase
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
+from flaskr.config import Config
 
 
 class Base(DeclarativeBase):
     pass
 
 
-app = Flask(__name__)
-app.config['SECRET_KEY'] = 'f8db447af15aaeb8af3f4a0d347bef8a' # env. variable
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///site.db'
-bcrypt = Bcrypt(app)        # Initialization
-login_manager = LoginManager(app)
-login_manager.login_view = 'login'
+bcrypt = Bcrypt()        # Initialization
+login_manager = LoginManager()
+login_manager.login_view = 'users.login'
 login_manager.login_message_category = 'info'
+db = SQLAlchemy()    # Initialization
 
-with app.app_context():
-    db = SQLAlchemy(app=app, model_class=Base)    # Initialization
 
-from flaskr import routes
+def create_app(config_class=Config):
+    app = Flask(__name__)
+    app.config.from_object(config_class)
+
+    db.init_app(app)
+    bcrypt.init_app(app)
+    login_manager.init_app(app)
+
+    from flaskr.users.routes import users
+    from flaskr.posts.routes import posts
+    from flaskr.main.routes import main
+
+    app.register_blueprint(users)
+    app.register_blueprint(posts)
+    app.register_blueprint(main)
+
+    return app
