+++ flaskr/templates/login.html
@@ -49,7 +49,7 @@
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Need an Account? <a class="ms-2 article-title" href="{{ url_for('register') }}">Sign Up Now</a>
+            Need an Account? <a class="ms-2 article-title" href="{{ url_for('users.register') }}">Sign Up Now</a>
         </small>
     </div>
 {% endblock content %}
