+++ flaskr/templates/register.html
@@ -70,7 +70,7 @@
     </div>
     <div class="border-top pt-3">
         <small class="text-muted">
-            Already Have an Account? <a class="ms-2 article-title" href="{{ url_for('login') }}">Sign In</a>
+            Already Have an Account? <a class="ms-2 article-title" href="{{ url_for('users.login') }}">Sign In</a>
         </small>
     </div>
 {% endblock content %}
