+++ run.py
@@ -1,5 +1,6 @@
-from flaskr import app
+from flaskr import create_app
 
+app = create_app()
 
 if __name__ == '__main__':
     app.run(debug=True)
